+++ modules/validators/src/include/Validators.hpp
@@ -93,5 +93,8 @@ mustBeNonSparse(const ArrayOfVector& args, int argPosition);
 NLSVALIDATORS_IMPEXP void
 mustBeReal(const ArrayOfVector& args, int argPosition);
 //=============================================================================
+NLSVALIDATORS_IMPEXP void
+mustBeInteger(const ArrayOfVector& args, int argPosition);
+//=============================================================================
 }
 //=============================================================================